ml_service/requirements.txt
praat-parselmouth==0.4.3
numpy==1.24.3
scipy==1.11.3
nolds==0.5.2
pydub==0.25.1
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.0.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
